requirements.txt
Flask>=2.3
gspread
google-auth
